// courseMangement/backend/src/schemas/courseSchema.ts
import { z } from "zod";

export const CourseZodSchema = z.object({
	courseId: z.string(),
	title: z.string(),
	courseCode: z.string().optional(),
	universityCode: z.string().optional(),
	universityName: z.string().optional(),
	department: z.string().optional(),
	category: z.string().optional(),
	specialization: z.string().optional(),
	courseLevel: z.string().optional(),
	description: z.string().optional(),
	summary: z.string().optional(),
	prerequisites: z.array(z.string()).default([]),
	learningOutcomes: z.array(z.string()).default([]),
	teachingMethodology: z.string().optional(),
	assessmentMethods: z.array(z.string()).default([]),
	credits: z.number().default(0),
	duration: z.number().default(0),
	language: z.string().optional(),
	syllabusUrl: z.string().optional(),
	keywords: z.array(z.string()).default([]),
	instructor: z.string().optional(),
	instructorEmail: z.string().optional(),
	officeLocation: z.string().optional(),
	intake: z.string().optional(),
	admissionOpenYears: z.string().optional(),
	attendanceType: z.string().optional(),
	tuitionFeeFirstYear: z.number().default(0),
	tuitionFeeTotal: z.number().default(0),
	currency: z.string().optional(),
	applicationFee: z.number().default(0),
	applicationFeeCurrency: z.string().optional(),
	applicationFeeWaived: z.boolean().default(false),
	requiredMaterials: z.string().optional(),
	minIELTS: z.number().default(0),
	minTOEFL: z.number().default(0),
	minPTE: z.number().default(0),
	minDuolingo: z.number().default(0),
	minCambridge: z.number().default(0),
	greRequired: z.boolean().default(false),
	greScore: z.number().default(0),
	gmatRequired: z.boolean().default(false),
	gmatScore: z.number().default(0),
	satRequired: z.boolean().default(false),
	satScore: z.number().default(0),
	actRequired: z.boolean().default(false),
	actScore: z.number().default(0),
	acceptanceRate: z.number().default(0),
	courseUrl: z.string().optional(),
});

export type ICourse = z.infer<typeof CourseZodSchema>;